audit_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
slack-sdk
notion-client
scikit-learn
pandas